feat: Add mock implementation library
This commit introduces a new library module named `mock` which provides tools for mocking MviModels.
It includes `MockMviModel` and `MockMviModelWithActions` classes along with their respective processor types and a utility function to create a dedicated coroutine scope.
The `build.gradle.kts` is configured for publishing, with the library described for clarity.
`.gitignore` file is also added.
Also modified `gradle.xml` and `settings.gradle.kts` to add new module to the project.