refactor: Rename mock MVI model classes and improve their integration with ViewModel
This commit renames `MockMviModel` to `MockMviViewModel` and `MockMviModelWithEvents` to `MockMviViewModelWithEvents`, and renames `MviModelScope` to `MockMviViewModelScope` within the `mock` module.
It also introduces `ProcessorScope` and improves the integration of the mock MVI model classes with the Android ViewModel, by making them inherit from ViewModel.
Additionally, it adds the `androidx.lifecycle.viewmodel` library as a dependency.